Refactor Windows window management
- Introduced `BasicWindowProc` for improved Windows window management.
- Updated `WindowsWindowUtils` to handle window procedures and dispose them correctly.
- Replaced `TitleBarWindowProc` with `CustomTitleBarWindowProc` and removed related methods.
- Added `accentColor` state flow for dynamically updating the window's accent color.
- Updated `WindowsWindowFrame` to use the new window procedures.
- Refactor `AppTheme.desktop.kt` to get the current window's accent color.
