+++ app/shared/ui-foundation/src/desktopMain/kotlin/ui/foundation/theme/AppTheme.desktop.kt
@@ -12,16 +12,13 @@ package me.him188.ani.app.ui.foundation.theme
 import androidx.compose.material3.ColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.snapshotFlow
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.takeOrElse
 import com.materialkolor.PaletteStyle
 import com.materialkolor.dynamicColorScheme
-import me.him188.ani.app.platform.window.WindowsWindowUtils
 import me.him188.ani.app.ui.foundation.LocalPlatform
 import me.him188.ani.app.ui.foundation.layout.LocalPlatformWindow
-import me.him188.ani.utils.coroutines.combine
 import me.him188.ani.utils.platform.isWindows
 
 @Composable
@@ -31,12 +28,15 @@ actual fun appColorScheme(
     useBlackBackground: Boolean,
     isDark: Boolean,
 ): ColorScheme {
-    return DesktopDynamicTheme(
-        useDynamicTheme = useDynamicTheme,
-        useBlackBackground = useBlackBackground,
-        isDark = isDark,
-    ) ?: dynamicColorScheme(
-        primary = seedColor,
+    val actualSeedColor = if (useDynamicTheme && isPlatformSupportDynamicTheme()) {
+        val currentWindowColor by LocalPlatformWindow.current.accentColor.collectAsState()
+        //fallback to default color, if platform can't get the accent color
+        currentWindowColor.takeOrElse { seedColor }
+    } else {
+        seedColor
+    }
+    return dynamicColorScheme(
+        primary = actualSeedColor,
         isDark = isDark,
         isAmoled = useBlackBackground,
         style = PaletteStyle.TonalSpot,
@@ -47,52 +47,6 @@ actual fun appColorScheme(
 }
 
 @Composable
-private fun DesktopDynamicTheme(
-    useDynamicTheme: Boolean,
-    useBlackBackground: Boolean,
-    isDark: Boolean,
-): ColorScheme? {
-    return when {
-        !useDynamicTheme -> null
-        LocalPlatform.current.isWindows() -> {
-            val platformWindow = LocalPlatformWindow.current
-            val isDarkState = rememberUpdatedState(isDark)
-            val useBlackBackgroundState = rememberUpdatedState(useBlackBackground)
-            remember(platformWindow, isDarkState, useBlackBackgroundState) {
-                WindowsWindowUtils.instance.windowAccentColor(platformWindow)
-                    .combine(
-                        snapshotFlow { isDarkState.value },
-                        snapshotFlow { useBlackBackgroundState.value },
-                    ) { color, isDark, useBlackBackground ->
-                        if (color != Color.Unspecified) {
-                            dynamicColorScheme(
-                                primary = color,
-                                isDark = isDark,
-                                isAmoled = useBlackBackground,
-                                style = PaletteStyle.TonalSpot,
-                                modifyColorScheme = { colorScheme ->
-                                    modifyColorSchemeForBlackBackground(colorScheme, isDark, useBlackBackground)
-                                },
-                            )
-                        } else {
-                            null
-                        }
-                    }
-            }.collectAsState(null).value
-        }
-
-        else -> null
-    }
-}
-
-@Composable
 actual fun isPlatformSupportDynamicTheme(): Boolean {
-    return if (LocalPlatform.current.isWindows()) {
-        val platformWindow = LocalPlatformWindow.current
-        remember(platformWindow) {
-            WindowsWindowUtils.instance.isExtendToTitleBar(platformWindow)
-        }.collectAsState(false).value
-    } else {
-        false
-    }
+    return LocalPlatform.current.isWindows()
 }
